Refactor and cleanup;
- remove translations relating to index calculations being locked
- rename and reorganize files related to hankeform
