+++ src/domain/hanke/edit/components/FormStepIndicator.module.scss
