+++ src/domain/hanke/edit/utils.ts
@@ -31,12 +31,9 @@ export const convertHankeDataToFormState = (hankeData: HankeDataDraft): HankeDat
   toteuttajat: hankeData.toteuttajat ? hankeData.toteuttajat : [],
 });
 
-export const isHankeEditingDisabled = ({ alkuPvm, tilat }: HankeDataDraft | HankeDataFormState) => {
+export const isHankeEditingDisabled = ({ alkuPvm }: HankeDataDraft | HankeDataFormState) => {
   if (alkuPvm && isBefore(parseISO(alkuPvm), new Date())) {
     return 'STARTED';
-  }
-  if (tilat && tilat.onLiikenneHaittaIndeksi) {
-    return 'INDEX_CALCULATED';
   }
 
   return false;
