+++ src/domain/map/components/HankeSidebar/HankeIndexes.module.scss
@@ -21,6 +21,10 @@
   display: flex;
 }
 
+.indexInfo {
+  padding: 0.4rem 0.2rem 0.4rem 1rem;
+}
+
 .indexTooltip {
   padding-left: 1rem;
 }
